Increase devenv timeout to 60 mins
Devenv had a smaller timeout (30 min) than the 60 min timeout
for the whole stage. Remove the smaller timeout and allow the
outer timeout to take precedence.

The proper fix here is to add a local cache of the docker
images near the CI workers, preventing the lengthly container
image download. However, this will do for now, as timeouts in
CI are less to do with enforcing performance goals, and are
instead about making sure a CI job does not consume a CI
worker forever.